+++ CHANGELOG.md
@@ -5,7 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [vNext]
+
+## [0.5.0] / 2019-06-25
+- Changed `Nuke.Common` version to 0.20.1
 - Added resolving of string secrets from parameters
+- Fixed getting the value of `KeyVaultSettingsAttribute`
+
 ## [0.4.3] / 2019-02-23
 - Changed `Nuke.Common` version to 0.17.0
 
@@ -29,7 +34,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.0] / 2018-05-22
 - Initial release
 
-[vNext]: https://github.com/nuke-build/azure-keyvault/compare/0.4.3...HEAD
+[vNext]: https://github.com/nuke-build/azure-keyvault/compare/0.5.0...HEAD
+[0.5.0]: https://github.com/nuke-build/azure-keyvault/compare/0.4.3...0.5.0
 [0.4.3]: https://github.com/nuke-build/azure-keyvault/compare/0.4.1...0.4.3
 [0.4.1]: https://github.com/nuke-build/azure-keyvault/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/nuke-build/azure-keyvault/compare/0.3.0...0.4.0